Avoid Mobile Wikipedia on Desktop internet
Mobile wikipedia is a nice read as it is cleaner by default,
but uses more JS and so on. As the content is the same, we just
redirect to the non-mobile Wikipedia. That could easily be reversed,
so that e.g. reddit communities wouldn't be littered by bots posting
this link (or the other) in comments all the time…